role permission backend

+++ app/Http/Controllers/Admin/Role/RolePermissionController.php
@@ -1,16 +1,13 @@
 <?php
-
 namespace App\Http\Controllers\Admin\Role;
 
-use Auth;
-use App\Models\User;
-use App\Models\Role\Role;
-use Illuminate\Http\Request;
-use Yajra\DataTables\DataTables;
-use Illuminate\Support\Facades\DB;
-use App\Models\Role\RolePermission;
-use App\Http\Controllers\Controller;
 use App\Http\Controllers\BaseController;
+use App\Models\Role\Role;
+use App\Models\Role\RolePermission;
+use App\Models\User;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Yajra\DataTables\DataTables;
 
 class RolePermissionController extends BaseController
 {
@@ -19,9 +16,9 @@ class RolePermissionController extends BaseController
      */
     public function index()
     {
-        $data= DB::table('roles as r')
-        ->join('role_permissions as rp','r.id','rp.role_id')
-        ->selectRaw('r.name,r.id as idd,count(rp.id) total_perms,r.created_at,r.status,r.updated_at,f_username(r.updated_by) updated_by,f_username(r.created_by) created_by')->groupBy('r.id','r.name','r.created_at','r.status','r.updated_at','r.updated_by','r.created_by')->get();
+        $data = DB::table('roles as r')
+            ->join('role_permissions as rp', 'r.id', 'rp.role_id')
+            ->selectRaw('r.name,r.id as idd,count(rp.id) total_perms,r.created_at,r.status,r.updated_at,f_username(r.updated_by) updated_by,f_username(r.created_by) created_by')->groupBy('r.id', 'r.name', 'r.created_at', 'r.status', 'r.updated_at', 'r.updated_by', 'r.created_by')->get();
 
         return DataTables::of($data)->addIndexColumn()->make(true);
 
@@ -51,7 +48,7 @@ class RolePermissionController extends BaseController
     public function roleSave(Request $request)
     {
 
-        $auth = User::where('email',$request->useEmail)->first();
+        $auth = User::where('email', $request->useEmail)->first();
 
         $validator = validator($request->all(),
             ['roleName' => 'required|unique:roles,name'],
@@ -60,23 +57,23 @@ class RolePermissionController extends BaseController
             return response()->json(['message' => $validator->errors()->all(), 'types' => 'e']);
         }
 
-        if(sizeof($request->checkedNames)==0){
+        if (sizeof($request->checkedNames) == 0) {
             return response()->json(['message' => 'No permission assigned.', 'types' => 'e']);
 
         }
 
-        $role =  new Role;
-        $role->name =  $request->roleName;
-        $role->display_name =  $request->roleName;
-        $role->status =  1;
-        $role->created_by =  $auth->id;
+        $role               = new Role;
+        $role->name         = $request->roleName;
+        $role->display_name = $request->roleName;
+        $role->status       = 1;
+        $role->created_by   = $auth->id;
         $role->save();
 
-        foreach($request->checkedNames as $perm){
-            $perms = new RolePermission;
-            $perms->role_id = $role->id;
+        foreach ($request->checkedNames as $perm) {
+            $perms               = new RolePermission;
+            $perms->role_id      = $role->id;
             $perms->feature_name = $perm;
-            $perms->created_by =  $auth->id;
+            $perms->created_by   = $auth->id;
 
             $perms->save();
         }
@@ -105,8 +102,8 @@ class RolePermissionController extends BaseController
 
     public function getPermissionList(Request $request)
     {
-        $data= DB::table('role_permissions as r')->where('role_id',$request->id)
-        ->selectRaw('r.feature_name')->get();
+        $data = DB::table('role_permissions as r')->where('role_id', $request->id)
+            ->selectRaw('r.feature_name')->get();
         return response()->json($data);
     }
     /**
@@ -114,16 +111,13 @@ class RolePermissionController extends BaseController
      */
     public function update(Request $request)
     {
-        // dd($request->all());
-
-        $role = Role::find($request->role_id);
-        $role->name     = $request->roleName ? $request->roleName : $role->name;
-        $role->status   = $request->status_val ? $request->status_val : $role->status;
+        $role         = Role::find($request->role_id);
+        $role->name   = $request->roleName ? $request->roleName : $role->name;
+        $role->status = $request->status_val ? $request->status_val : $role->status;
 
         return response()->json(['message' => 'Successfully Role Saved.', 'types' => 's']);
 
     }
-
 
     /**
      * Remove the specified resource from storage.
@@ -151,16 +145,18 @@ class RolePermissionController extends BaseController
     }
     public function destroy(Request $request)
     {
+        $this->authorize('role_delete');
+        // dd($this->authorize('role_delete'));
         if ($request->id) {
 
-            $role = Role::where('id', $request->id)->first();
-            $perms = RolePermission::where('role_id',$request->id)->get();
-            foreach($perms as $perm){
+            $role  = Role::where('id', $request->id)->first();
+            $perms = RolePermission::where('role_id', $request->id)->get();
+            foreach ($perms as $perm) {
                 $perm->delete();
             }
             $role->delete();
             $success = '';
-            return $this->SuccessResponse($success, 'Successfully Department deleted.');
+            return $this->SuccessResponse($success, 'Successfully Role deleted.');
 
         } else {
             $error = 'Id can not be null.';
